Reorganize repository structure
