+++ index.html
@@ -6,13 +6,13 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>慈悲なきアイオニア　非公式キャラクターシート</title>
     <script src="https://unpkg.com/vue@3/dist/vue.global.js"></script>
-    <link rel="shortcut icon" type="image/x-icon" href="favicon.ico">
+    <link rel="shortcut icon" type="image/x-icon" href="public/favicon.ico">
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link
         href="https://fonts.googleapis.com/css2?family=Cinzel+Decorative:wght@700&family=Noto+Serif+JP:wght@400;700&family=Noto+Sans+JP:wght@400;700&display=swap"
         rel="stylesheet">
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="public/style.css">
 </head>
 
 <body>
@@ -333,11 +333,11 @@
         </div>
 
     </div>
-    <script src="cocofoliaExporter.js"></script>
-    <script src="gameData.js"></script>
-    <script src="utils.js"></script>
-    <script src="dataManager.js"></script>
-    <script src="main.js"></script>
+    <script src="src/cocofoliaExporter.js"></script>
+    <script src="src/gameData.js"></script>
+    <script src="src/utils.js"></script>
+    <script src="src/dataManager.js"></script>
+    <script src="src/main.js"></script>
 </body>
 
 </html>
