+++ package.json
@@ -2,14 +2,14 @@
   "name": "aioniacs",
   "version": "1.0.0",
   "description": "慈悲なきアイオニアのための非公式キャラクターシート作成ツールです",
-  "main": "main.js",
+  "main": "src/main.js",
   "scripts": {
     "lint": "npm run lint:js && npm run lint:css && npm run lint:html",
-    "lint:js": "eslint .",
-    "lint:css": "stylelint \"**/*.css\"",
-    "lint:html": "htmlhint \"**/*.html\"",
+    "lint:js": "eslint --config ./config/eslint.config.js .",
+    "lint:css": "stylelint \"**/*.css\" --config ./config/.stylelintrc.json",
+    "lint:html": "htmlhint \"**/*.html\" --config ./config/.htmlhintrc",
     "test": "jest",
-    "e2e": "playwright test",
+    "e2e": "playwright test --config ./config/playwright.config.js",
     "pree2e": "playwright install"
   },
   "keywords": [],
@@ -26,6 +26,8 @@
     "stylelint-config-standard": "^38.0.0"
   },
   "jest": {
-    "testPathIgnorePatterns": ["<rootDir>/e2e/"]
+    "testPathIgnorePatterns": [
+      "<rootDir>/tests/e2e/"
+    ]
   }
 }
